+++ .setup/cpp/simStateSpace-solve-syl-dot.cpp
@@ -6,7 +6,7 @@
 #include <RcppArmadillo.h>
 // [[Rcpp::depends(RcppArmadillo)]]
 // [[Rcpp::export(.SolveSyl)]]
-arma::mat SolveSyl(arma::mat A, arma::mat B, arma::mat C) {
+arma::mat SolveSyl(const arma::mat A, const arma::mat B, const arma::mat C) {
   arma::mat X;
   arma::syl(X, A, B, C);
   return X;