+++ .setup/cpp/simStateSpace-ssm-cov-eta-dot.cpp
@@ -1,0 +1,15 @@
+// -----------------------------------------------------------------------------
+// edit .setup/cpp/simStateSpace-ssm-cov-eta-dot.cpp
+// Ivan Jacob Agaloos Pesigan
+// -----------------------------------------------------------------------------
+
+#include <RcppArmadillo.h>
+// [[Rcpp::depends(RcppArmadillo)]]
+// [[Rcpp::export(.SSMCovEta)]]
+arma::mat SSMCovEta(const arma::mat& beta, const arma::mat& psi) {
+  arma::vec vec_sigma = arma::solve(
+      arma::eye(beta.n_rows * beta.n_rows, beta.n_rows * beta.n_rows) -
+          arma::kron(beta, beta),
+      arma::vectorise(psi));
+  return arma::reshape(vec_sigma, beta.n_rows, beta.n_rows);
+}