+++ .sim/make.sh
@@ -8,14 +8,7 @@
 
 # Define an array of directories
 directories=(
-  "/scratch/ibp5092/betaDelta"
-  "/scratch/ibp5092/betaMC"
-  "/scratch/ibp5092/betaNB"
-  "/scratch/ibp5092/betaSandwich"
-  "/scratch/ibp5092/semmcci"
   "/scratch/ibp5092/simStateSpace"
-  "/scratch/ibp5092/bootStateSpace"
-  "/scratch/ibp5092/cTMed"
 )
 
 # SIF path
@@ -27,7 +20,7 @@ for dir in "${directories[@]}"; do
     echo "Processing: $dir"
     cd "$dir" || exit
     apptainer exec "$sif_path" make all
-    # apptainer exec "$sif_path" make auto
+    apptainer exec "$sif_path" make auto
   fi
 done